==> 1 Readmission Dataset.Rmd <==
---
title: "INFR 6197 & INFR 6198"
output: html_notebook
---
Dataset SQL

```{r}
library(bigrquery)

# This query represents dataset "Surgical Patients 112121.1" for domain "person" and was generated for All of Us Registered Tier Dataset v5
dataset_person_sql <- paste("
    SELECT
        person.SEX_AT_BIRTH_CONCEPT_ID,
        person.GENDER_CONCEPT_ID,
        person.BIRTH_DATETIME as DATE_OF_BIRTH,
        person.ETHNICITY_CONCEPT_ID,
        person.PERSON_ID,
        person.RACE_CONCEPT_ID,
        p_race_concept.concept_name as RACE,
        p_gender_concept.concept_name as GENDER,
        p_ethnicity_concept.concept_name as ETHNICITY,
        p_sex_at_birth_concept.concept_name as SEX_AT_BIRTH 
    FROM
        `person` person 
    LEFT JOIN
        `concept` p_race_concept 
            on person.race_concept_id = p_race_concept.CONCEPT_ID 
    LEFT JOIN
        `concept` p_gender_concept 
            on person.gender_concept_id = p_gender_concept.CONCEPT_ID 
    LEFT JOIN
        `concept` p_ethnicity_concept 
            on person.ethnicity_concept_id = p_ethnicity_concept.CONCEPT_ID 
    LEFT JOIN
        `concept` p_sex_at_birth_concept 
            on person.sex_at_birth_concept_id = p_sex_at_birth_concept.CONCEPT_ID  
    WHERE
        person.PERSON_ID IN (
            SELECT
                person_id  
            FROM
                `cb_search_person` cb_search_person  
            WHERE
                cb_search_person.person_id IN (
                    SELECT
                        criteria.person_id 
                    FROM
                        (SELECT
                            DISTINCT person_id,
                            entry_date,
                            concept_id 
                        FROM
                            `cb_search_all_events` 
                        WHERE
                            (
                                is_standard = 1 
                                AND concept_id IN (2108427, 2109451, 2108972, 2109385, 2109418, 2109071, 2109005, 2110128, 2109121, 2109154, 2109435, 2108956, 2109369, 2109402, 2109055, 2109567, 2108989, 2109501, 2108510, 2109022, 2109336, 2109004, 2109417, 2108426, 2109450, 2109136, 2109219, 2107716, 927092, 2109153, 2109384, 2109318, 2110375, 2109467, 2108988, 2109500, 2109401, 2109087, 2109120, 2109021, 2109054, 2109137, 2109335, 2109368, 2109003, 2109135, 2109069, 2109218, 2109383, 2109317, 40757103, 2108475, 2108508, 2109020, 2109433, 2108954, 2109119, 2109152, 2109086, 2107732, 725062, 2109400, 2109334, 2110391, 2109068, 2109002, 2102412, 2109316, 2109415, 2110373, 2109382, 2108507, 2109052, 2108953, 2109151, 2109085, 2109118, 2110390, 725063, 2109333, 2109366, 2109067, 2109001, 709956, 2109034, 2109315, 2108423, 2109447, 2108968, 2109381, 2109018, 709973, 46257444, 2109216, 2101916, 2109117, 2109332, 2109431, 2109365, 2109398, 2109066, 2108207, 2109347, 2109314, 2108967, 40756846, 2108488, 2109000, 2109446, 2109083, 2109017, 709974, 2101915, 2109331, 2109496, 2109430, 2109164, 2109065, 2108206, 2109313, 2109346, 2109445, 2108966, 2109148, 2109082, 2101914, 2109214, 2108917, 2109429, 2109444, 2109378, 2108965, 2108998, 2109081, 2109395, 2109428, 2109048, 2108470, 2109129, 2109162, 2108964, 2109377, 2109410, 2109063, 2102820, 2108997, 2109030, 2109146, 2109311, 2109344, 40756949, 2109460, 2109394, 2109047, 2109080, 2109161, 2109392, 2108963, 2109409, 2108930, 2109128, 2102819, 2109657, 40756884, 2109343, 2109376, 2108980, 2109079, 2108501, 2109525, 2109046, 2102735, 709958, 2102537, 2108995, 2109028, 2108962, 2102818, 2109127, 2109160, 2109606, 2102521, 2108384, 2109408, 2109144, 2108021, 2109045, 2109078, 40757063, 2109390, 2109027, 2109060, 2108961, 2109126, 2104799, 2109209, 2108928, 2110365, 2109044, 2108978, 2109143, 2109389, 2109059, 40757002, 2108993, 2109026, 40757068, 40757134, 2109125, 2107738, 2108960, 2108861, 2109406, 2109043, 2102552, 2102519, 2109010, 2109208, 2109388, 46257749, 2108975, 2109008, 2109421, 2109454, 2109124, 2109058, 2109157, 2109339, 2109372, 2108959, 2109405, 2109075, 2109009, 2109042, 2109207, 2109141, 2109387, 2109453, 2108974, 2109123, 2109057, 2109371, 2109404, 2108991, 2109024, 2109437, 42742541, 2109041, 2109206, 2109419, 2109386, 2109072, 2109155, 40757001, 2109122, 2109403, 2110394, 2109370, 2109023, 2109056, 2108957, 2108990, 2109502, 4185284, 2109073) 
                            )) criteria ) 
                        AND cb_search_person.person_id IN (SELECT
                            criteria.person_id 
                        FROM
                            (SELECT
                                DISTINCT person_id,
                                entry_date,
                                concept_id 
                            FROM
                                `cb_search_all_events` 
                            WHERE
                                (
                                    is_standard = 1 
                                    AND concept_id IN (9201, 8717, 262) 
                                )) criteria ) )", sep="")

dataset_person_df <- bq_table_download(bq_dataset_query(Sys.getenv("WORKSPACE_CDR"), dataset_person_sql, billing=Sys.getenv("GOOGLE_PROJECT")), bigint="integer64")

# Generate the csv file that other workbooks will process
write.csv(dataset_person_df, "person.csv")

# Create a unique list of all inpatient visits for patients with a surgical encounter
person_id <- dataset_person_df %>% drop_na(PERSON_ID)
person_id <- unique(person_id$PERSON_ID)

# Custom query for all inpatient visits that had a surgical encounter
# Paste the person ids from the surgical population query
dataset_inpatient_sql <- paste("
    SELECT
        person_id,
        visit_occurrence_id,
        entry_date,
        entry_datetime
    FROM
        cb_search_all_events
    WHERE
        concept_id IN (9201, 8717, 262) 
        AND is_standard = 1
        AND person_id in (", paste0(person_id, collapse=","),")", sep="")

dataset_inpatient_sql <- bq_table_download(bq_dataset_query(Sys.getenv("WORKSPACE_CDR"), dataset_inpatient_sql, billing=Sys.getenv("GOOGLE_PROJECT")), bigint="integer64")

write.csv(dataset_inpatient_sql, "inpatients.csv")

# This query represents dataset "Surgical Patients 112121.1" for domain "survey" and was generated for All of Us Registered Tier Dataset v5
dataset_84280434_survey_sql <- paste("
    SELECT
        answer.answer,
        answer.question,
        answer.person_id,
        answer.survey_version_name,
        answer.survey_datetime,
        answer.answer_concept_id,
        answer.survey_version_concept_id,
        answer.survey,
        answer.question_concept_id  
    FROM
        `ds_survey` answer   
    WHERE
        (
            question_concept_id IN (
                1585772, 1585778, 1585940, 1585766
            )
        )  
        AND (
            answer.PERSON_ID IN (
                SELECT
                    person_id  
                FROM
                    `cb_search_person` cb_search_person  
                WHERE
                    cb_search_person.person_id IN (
                        SELECT
                            criteria.person_id 
                        FROM
                            (SELECT
                                DISTINCT person_id,
                                entry_date,
                                concept_id 
                            FROM
                                `cb_search_all_events` 
                            WHERE
                                (
                                    is_standard = 1 
                                    AND concept_id IN (2108427, 2109451, 2108972, 2109385, 2109418, 2109071, 2109005, 2110128, 2109121, 2109154, 2109435, 2108956, 2109369, 2109402, 2109055, 2109567, 2108989, 2109501, 2108510, 2109022, 2109336, 2109004, 2109417, 2108426, 2109450, 2109136, 2109219, 2107716, 927092, 2109153, 2109384, 2109318, 2110375, 2109467, 2108988, 2109500, 2109401, 2109087, 2109120, 2109021, 2109054, 2109137, 2109335, 2109368, 2109003, 2109135, 2109069, 2109218, 2109383, 2109317, 40757103, 2108475, 2108508, 2109020, 2109433, 2108954, 2109119, 2109152, 2109086, 2107732, 725062, 2109400, 2109334, 2110391, 2109068, 2109002, 2102412, 2109316, 2109415, 2110373, 2109382, 2108507, 2109052, 2108953, 2109151, 2109085, 2109118, 2110390, 725063, 2109333, 2109366, 2109067, 2109001, 709956, 2109034, 2109315, 2108423, 2109447, 2108968, 2109381, 2109018, 709973, 46257444, 2109216, 2101916, 2109117, 2109332, 2109431, 2109365, 2109398, 2109066, 2108207, 2109347, 2109314, 2108967, 40756846, 2108488, 2109000, 2109446, 2109083, 2109017, 709974, 2101915, 2109331, 2109496, 2109430, 2109164, 2109065, 2108206, 2109313, 2109346, 2109445, 2108966, 2109148, 2109082, 2101914, 2109214, 2108917, 2109429, 2109444, 2109378, 2108965, 2108998, 2109081, 2109395, 2109428, 2109048, 2108470, 2109129, 2109162, 2108964, 2109377, 2109410, 2109063, 2102820, 2108997, 2109030, 2109146, 2109311, 2109344, 40756949, 2109460, 2109394, 2109047, 2109080, 2109161, 2109392, 2108963, 2109409, 2108930, 2109128, 2102819, 2109657, 40756884, 2109343, 2109376, 2108980, 2109079, 2108501, 2109525, 2109046, 2102735, 709958, 2102537, 2108995, 2109028, 2108962, 2102818, 2109127, 2109160, 2109606, 2102521, 2108384, 2109408, 2109144, 2108021, 2109045, 2109078, 40757063, 2109390, 2109027, 2109060, 2108961, 2109126, 2104799, 2109209, 2108928, 2110365, 2109044, 2108978, 2109143, 2109389, 2109059, 40757002, 2108993, 2109026, 40757068, 40757134, 2109125, 2107738, 2108960, 2108861, 2109406, 2109043, 2102552, 2102519, 2109010, 2109208, 2109388, 46257749, 2108975, 2109008, 2109421, 2109454, 2109124, 2109058, 2109157, 2109339, 2109372, 2108959, 2109405, 2109075, 2109009, 2109042, 2109207, 2109141, 2109387, 2109453, 2108974, 2109123, 2109057, 2109371, 2109404, 2108991, 2109024, 2109437, 42742541, 2109041, 2109206, 2109419, 2109386, 2109072, 2109155, 40757001, 2109122, 2109403, 2110394, 2109370, 2109023, 2109056, 2108957, 2108990, 2109502, 4185284, 2109073) 
                                )) criteria ) 
                            AND cb_search_person.person_id IN (SELECT
                                criteria.person_id 
                            FROM
                                (SELECT
                                    DISTINCT person_id,
                                    entry_date,
                                    concept_id 
                                FROM
                                    `cb_search_all_events` 
                                WHERE
                                    (
                                        is_standard = 1 
                                        AND concept_id IN (9201, 8717, 262) 
                                    )) criteria ) ))", sep="")

dataset_84280434_survey_df <- bq_table_download(bq_dataset_query(Sys.getenv("WORKSPACE_CDR"), dataset_84280434_survey_sql, billing=Sys.getenv("GOOGLE_PROJECT")), bigint="integer64")

# Generate the csv file that other workbooks will process
write.csv(dataset_84280434_survey_df, "survey.csv")

# This query represents dataset "Surgical Patients 112121.1" for domain "procedure" and was generated for All of Us Registered Tier Dataset v5
dataset_50631304_procedure_sql <- paste("
    SELECT
        procedure.person_id,
        procedure.procedure_concept_id,
        p_standard_concept.concept_name as standard_concept_name,
        p_standard_concept.concept_code as standard_concept_code,
        p_standard_concept.vocabulary_id as standard_vocabulary,
        procedure.procedure_datetime,
        procedure.procedure_type_concept_id,
        p_type.concept_name as procedure_type_concept_name,
        procedure.modifier_concept_id,
        p_modifier.concept_name as modifier_concept_name,
        procedure.quantity,
        procedure.visit_occurrence_id,
        p_visit.concept_name as visit_occurrence_concept_name,
        procedure.procedure_source_value,
        procedure.procedure_source_concept_id,
        p_source_concept.concept_name as source_concept_name,
        p_source_concept.concept_code as source_concept_code,
        p_source_concept.vocabulary_id as source_vocabulary,
        procedure.qualifier_source_value 
    FROM
        ( SELECT
            * 
        FROM
            `procedure_occurrence` procedure 
        WHERE
            (
                procedure_concept_id IN  (
                    SELECT
                        DISTINCT c.concept_id 
                    FROM
                        `cb_criteria` c 
                    JOIN
                        (
                            select
                                cast(cr.id as string) as id 
                            FROM
                                `cb_criteria` cr 
                            WHERE
                                concept_id IN (
                                    2108427, 2109451, 2108972, 2109385, 2109418, 2109071, 2109005, 2110128, 2109121, 2109154, 2109435, 2108956, 2109369, 2109402, 2109055, 2109567, 2108989, 2109501, 2108510, 2109022, 2109336, 2109004, 2109417, 2108426, 2109450, 2109136, 2109219, 2107716, 927092, 2109153, 2109384, 2109318, 2110375, 2109467, 2108988, 2109500, 2109401, 2109087, 2109120, 2109021, 2109054, 2109137, 2109335, 2109368, 2109003, 2109135, 2109069, 2109218, 2109383, 2109317, 40757103, 2108475, 2108508, 2109020, 2109433, 2108954, 2109119, 2109152, 2109086, 2107732, 725062, 2109400, 2109334, 2110391, 2109068, 2109002, 2102412, 2109316, 2109415, 2110373, 2109382, 2108507, 2109052, 2108953, 2109151, 2109085, 2109118, 2110390, 725063, 2109333, 2109366, 2109067, 2109001, 709956, 2109034, 2109315, 2108423, 2109447, 2108968, 2109381, 2109018, 709973, 46257444, 2109216, 2101916, 2109117, 2109332, 2109431, 2109365, 2109398, 2109066, 2108207, 2109347, 2109314, 2108967, 40756846, 2108488, 2109000, 2109446, 2109083, 2109017, 709974, 2101915, 2109331, 2109496, 2109430, 2109164, 2109065, 2108206, 2109313, 2109346, 2109445, 2108966, 2109148, 2109082, 2101914, 2109214, 2108917, 2109429, 2109444, 2109378, 2108965, 2108998, 2109081, 2109395, 2109428, 2109048, 2108470, 2109129, 2109162, 2108964, 2109377, 2109410, 2109063, 2102820, 2108997, 2109030, 2109146, 2109311, 2109344, 40756949, 2109460, 2109394, 2109047, 2109080, 2109161, 2109392, 2108963, 2109409, 2108930, 2109128, 2102819, 2109657, 40756884, 2109343, 2109376, 2108980, 2109079, 2108501, 2109525, 2109046, 2102735, 709958, 2102537, 2108995, 2109028, 2108962, 2102818, 2109127, 2109160, 2109606, 2102521, 2108384, 2109408, 2109144, 2108021, 2109045, 2109078, 40757063, 2109390, 2109027, 2109060, 2108961, 2109126, 2104799, 2109209, 2108928, 2110365, 2109044, 2108978, 2109143, 2109389, 2109059, 40757002, 2108993, 2109026, 40757068, 40757134, 2109125, 2107738, 2108960, 2108861, 2109406, 2109043, 2102552, 2102519, 2109010, 2109208, 2109388, 46257749, 2108975, 2109008, 2109421, 2109454, 2109124, 2109058, 2109157, 2109339, 2109372, 2108959, 2109405, 2109075, 2109009, 2109042, 2109207, 2109141, 2109387, 2109453, 2108974, 2109123, 2109057, 2109371, 2109404, 2108991, 2109024, 2109437, 42742541, 2109041, 2109206, 2109419, 2109386, 2109072, 2109155, 40757001, 2109122, 2109403, 2110394, 2109370, 2109023, 2109056, 2108957, 2108990, 2109502, 4185284, 2109073
                                )
                        ) a 
                            ON (
                                c.path LIKE CONCAT('%.',
                            a.id,
                            '.%') 
                            OR c.path LIKE CONCAT('%.',
                            a.id) 
                            OR c.path LIKE CONCAT(a.id,
                            '.%') 
                            OR c.path = a.id) 
                        WHERE
                            is_standard = 1 
                            AND is_selectable = 1
                        )
                )  
                AND (
                    procedure.PERSON_ID IN (
                        SELECT
                            person_id  
                        FROM
                            `cb_search_person` cb_search_person  
                        WHERE
                            cb_search_person.person_id IN (
                                SELECT
                                    criteria.person_id 
                                FROM
                                    (SELECT
                                        DISTINCT person_id,
                                        entry_date,
                                        concept_id 
                                    FROM
                                        `cb_search_all_events` 
                                    WHERE
                                        (
                                            concept_id IN (9201, 8870, 8717, 262) 
                                            AND is_standard = 1 
                                        )) criteria ) 
                            )
                        )
                ) procedure 
            LEFT JOIN
                `concept` p_standard_concept 
                    ON procedure.procedure_concept_id = p_standard_concept.concept_id 
            LEFT JOIN
                `concept` p_type 
                    ON procedure.procedure_type_concept_id = p_type.concept_id 
            LEFT JOIN
                `concept` p_modifier 
                    ON procedure.modifier_concept_id = p_modifier.concept_id 
            LEFT JOIN
                `visit_occurrence` v 
                    ON procedure.visit_occurrence_id = v.visit_occurrence_id 
            LEFT JOIN
                `concept` p_visit 
                    ON v.visit_concept_id = p_visit.concept_id 
            LEFT JOIN
                `concept` p_source_concept 
                    ON procedure.procedure_source_concept_id = p_source_concept.concept_id", sep="")

# Formulate a Cloud Storage destination path for the data exported from BigQuery.
# NOTE: By default data exported multiple times on the same day will overwrite older copies.
#       But data exported on a different days will write to a new location so that historical
#       copies can be kept as the dataset definition is changed.
procedure_50631304_path <- file.path(
  Sys.getenv("WORKSPACE_BUCKET"),
  "bq_exports",
  Sys.getenv("OWNER_EMAIL"),
  strftime(lubridate::now(), "%Y%m%d"),  # Comment out this line if you want the export to always overwrite.
  "procedure_50631304",
  "procedure_50631304_*.csv")
message(str_glue('The data will be written to {procedure_50631304_path}. Use this path when reading ',
                 'the data into your notebooks in the future.'))

# Perform the query and export the dataset to Cloud Storage as CSV files.
# NOTE: You only need to run `bq_table_save` once. After that, you can
#       just read data from the CSVs in Cloud Storage.
bq_table_save(
  bq_dataset_query(Sys.getenv("WORKSPACE_CDR"), dataset_50631304_procedure_sql, billing = Sys.getenv("GOOGLE_PROJECT")),
  procedure_50631304_path,
  destination_format = "CSV")
# Read the data directly from Cloud Storage into memory.
# NOTE: Alternatively you can `gsutil -m cp {procedure_50631304_path}` to copy these files
#       to the Jupyter disk.
read_bq_export_from_workspace_bucket <- function(export_path) {
  col_types <- NULL
  bind_rows(
    map(system2('gsutil', args = c('ls', export_path), stdout = TRUE, stderr = TRUE),
        function(csv) {
          message(str_glue('Loading {csv}.'))
          chunk <- read_csv(pipe(str_glue('gsutil cat {csv}')), col_types = col_types, show_col_types = FALSE)
          if (is.null(col_types)) {
            col_types <- spec(chunk)
          }
          chunk
        }))
}
dataset_50631304_procedure_df <- read_bq_export_from_workspace_bucket(procedure_50631304_path)

dim(dataset_50631304_procedure_df)

write.csv(dataset_50631304_procedure_df, "procedure.csv")

# Create a unique list of visits from the surgical procedure query
visit_id <- dataset_50631304_procedure_df %>% drop_na(visit_occurrence_id)
visit_id <- unique(visit_id$visit_occurrence_id)

# Custom query for visit table information
# Paste the visit ids from the surgical population query
dataset_42670613_visit_sql <- paste("
    SELECT
        person_id,
        visit_occurrence_id,
        preceding_visit_occurrence_id,
        visit_start_date,
        visit_end_date,
        admitting_source_concept_id,
        provider_id,
        visit_concept_id
    FROM
        visit_occurrence
    WHERE
        visit_occurrence_id in (", paste0(visit_id, collapse=","),")", sep="")

dataset_42670613_visit_df <- bq_table_download(bq_dataset_query(Sys.getenv("WORKSPACE_CDR"), dataset_42670613_visit_sql, billing=Sys.getenv("GOOGLE_PROJECT")), bigint="integer64")

write.csv(dataset_42670613_visit_df, "visit.csv")
```
